added persistant volume
Add persistent volume for storage

This commit adds a persistent volume to the deployment configuration
file in order to support storage for the application. By adding the
persistent volume claim `flowise-pv-claim`, storage can be allocated to
`/root/.flowise` within the container. This change enables the
application to store and retrieve data across different container
restarts. The addition of the `disk-pv.yaml` file in the workflow allows
the persistent volume to be applied during deployment.